% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{add_effect_counts}
\alias{add_effect_counts}
\title{Add counts of levels of random effects to list of design matrices}
\usage{
add_effect_counts(design_matrices, candidate_formula, model_data)
}
\arguments{
\item{design_matrices}{list of design matrices}

\item{candidate_formula}{formula}

\item{model_data}{data.table}
}
\description{
Add counts of levels of random effects to list of design matrices
}
\keyword{internal}
