% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{add_independent_unit}
\alias{add_independent_unit}
\title{Add vector with independent unit to list of design matrices}
\usage{
add_independent_unit(design_matrices, model_data)
}
\arguments{
\item{design_matrices}{list of design matrices}

\item{model_data}{data.table}
}
\description{
Add vector with independent unit to list of design matrices
}
\keyword{internal}
