% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{add_protein_design}
\alias{add_protein_design}
\title{Add design matrix for proteins}
\usage{
add_protein_design(design_matrices, model_data)
}
\arguments{
\item{design_matrices}{list of design matrices}

\item{model_data}{data.table}
}
\description{
Add design matrix for proteins
}
\keyword{internal}
