% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{add_response}
\alias{add_response}
\title{Add response to list of design matrices}
\usage{
add_response(design_matrices, model_formula, model_data)
}
\arguments{
\item{design_matrices}{list of design matrices}

\item{model_formula}{formula}

\item{model_data}{data.table}
}
\description{
Add response to list of design matrices
}
\keyword{internal}
