% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoAPQModel.R
\name{getIsoRawFit}
\alias{getIsoRawFit}
\alias{getIsoRandomVariances}
\alias{getIsoProteinAbundances}
\alias{getIsoFixedEffects}
\alias{getIsoHessian}
\alias{getIsoLoglik}
\alias{getIsoFitStatus}
\title{Extract raw fit from IsoAPQModel object}
\usage{
getIsoRawFit(x)

getIsoRandomVariances(x)

getIsoProteinAbundances(x)

getIsoFixedEffects(x)

getIsoHessian(x)

getIsoLoglik(x)

getIsoFitStatus(x)
}
\arguments{
\item{x}{object of class IsoAPQModel}
}
\value{
list or other type returned by the model fitting function such as optim

numeric, vector of estimated variances of random effects

numeric, vector of values of estimated protein abundances

numeric, vector of values of estimated fixed effects

matrix - hessian of the negative log-likelihood evaluated at the solution

numeric- value of log-likelihood at the solution

character - message indicating if model converged
}
\description{
Extract raw fit from IsoAPQModel object

Extract variances of random effects from IsoAPQModel object

Extract protein abundances from IsoAPQModel object

Extract fixed effects from IsoAPQModel object

Extract hessian of the negative log-likelihood from IsoAPQModel object

Extract fixed effects from IsoAPQModel object

Extract fit status (message) from IsoAPQModel object
}
