% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoAPQModel.R
\name{getIsoModelDesign}
\alias{getIsoModelDesign}
\title{Extract model design from IsoAPQModel object}
\usage{
getIsoModelDesign(x)
}
\arguments{
\item{x}{object of class IsoAPQModel}
}
\value{
object of class IsoAPQModelDesign
}
\description{
Extract model design from IsoAPQModel object
}
