% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_fit.R
\name{fit_iso_fixed_model}
\alias{fit_iso_fixed_model}
\title{Fit a model based on model design}
\usage{
fit_iso_fixed_model(model_design, ...)
}
\arguments{
\item{model_design}{object of class IsoAPQModelDesign}
}
\value{
list
}
\description{
Fit a model based on model design
}
\keyword{internal}
