% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{get_candidate_formula}
\alias{get_candidate_formula}
\title{Get updated formula for use in lmer}
\usage{
get_candidate_formula(formula_chr)
}
\arguments{
\item{formula_chr}{as.character(formula)}
}
\description{
Get updated formula for use in lmer
}
\keyword{internal}
