% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{get_design_matrices}
\alias{get_design_matrices}
\title{Get design matrices}
\usage{
get_design_matrices(candidate_formula, has_random, has_fixed, model_data)
}
\arguments{
\item{candidate_formula}{formula}

\item{has_random}{if TRUE, there are random effects}

\item{has_fixed, }{if TRUE, there are fixed effects}

\item{model_data}{data.table}
}
\description{
Get design matrices
}
\keyword{internal}
