% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{get_empty_design}
\alias{get_empty_design}
\title{Design matrices for model with just proteins}
\usage{
get_empty_design(model_data)
}
\arguments{
\item{model_data}{data.table}
}
\description{
Design matrices for model with just proteins
}
\keyword{internal}
