% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{get_fixed_design}
\alias{get_fixed_design}
\title{Get design matrices for model with no random effects}
\usage{
get_fixed_design(candidate_formula, model_data)
}
\arguments{
\item{candidate_formula}{formula}

\item{model_data}{data.table}
}
\description{
Get design matrices for model with no random effects
}
\keyword{internal}
