% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{get_full_design}
\alias{get_full_design}
\title{Design matrices for full model}
\usage{
get_full_design(candidate_formula, model_data)
}
\arguments{
\item{candidate_formula}{formula}

\item{model_data}{data.table}
}
\description{
Design matrices for full model
}
\keyword{internal}
