% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_fit.R
\name{get_full_model_information}
\alias{get_full_model_information}
\title{Extract model information from optim() output and model design}
\usage{
get_full_model_information(fitted_model, model_design)

get_full_model_information(fitted_model, model_design)
}
\arguments{
\item{fitted_model}{output of optim()}

\item{model_design}{object of class IsoAPQModelDesign}
}
\value{
list

list
}
\description{
Extract model information from optim() output and model design

Extract model information from optim() output and model design
}
\keyword{internal}
