% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytical_gradient_fixed.R
\name{get_gradient_fixed}
\alias{get_gradient_fixed}
\title{Nonlinear mixed effects model based on isotopic distribution}
\usage{
get_gradient_fixed(
  proteins_design,
  fixed_design,
  response,
  num_fixed_effects,
  num_proteins
)
}
\arguments{
\item{proteins_design}{design matrix for protein parameters}

\item{fixed_design}{design matrix for fixed effects}

\item{response}{response vector}

\item{num_fixed_effects}{number of fixed effects}

\item{num_proteins}{number of proteins}

\item{random_effects_counts}{counts of groups for each random effect}

\item{independent_unit}{vector of values of the independent unit}
}
\value{
function of a single parameter that returns vector of gradient values
}
\description{
Nonlinear mixed effects model based on isotopic distribution
}
\keyword{internal}
