% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytical_gradient_full.R
\name{get_gradient_full_unit}
\alias{get_gradient_full_unit}
\title{Analytical gradient within a single independent unit}
\usage{
get_gradient_full_unit(
  random_design,
  proteins_design,
  fixed_design,
  response,
  num_random_effects,
  num_fixed_effects,
  num_proteins,
  num_rows,
  random_effects_counts
)
}
\arguments{
\item{random_design}{design matrix for random effects}

\item{proteins_design}{design matrix for protein parameters}

\item{fixed_design}{design matrix for fixed effects}

\item{response}{response vector}

\item{num_random_effects}{number of random effects}

\item{num_fixed_effects}{number of fixed effects}

\item{num_proteins}{number of proteins}

\item{num_rows}{number of observations}

\item{random_effects_counts}{counts of groups for each random effect}
}
\value{
function of a single parameter that returns analytical gradient
}
\description{
Analytical gradient within a single independent unit
}
\keyword{internal}
