% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_fit.R
\name{get_hessian}
\alias{get_hessian}
\title{Get hessian based on model design and parameters}
\usage{
get_hessian(model_design, pars)
}
\arguments{
\item{model_design}{object of class IsoAPQModelDesign}

\item{pars}{numeric}
}
\value{
matrix
}
\description{
Get hessian based on model design and parameters
}
\keyword{internal}
