% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analytical_gradient_random.R
\name{get_iso_random_gradient}
\alias{get_iso_random_gradient}
\title{Analytical gradient for the full model}
\usage{
get_iso_random_gradient(model_design)
}
\arguments{
\item{model_design}{object of class IsoAPQModelDesign}
}
\description{
Analytical gradient for the full model
}
\keyword{internal}
