% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_log_likelihood_random.R
\name{get_iso_random_loglikelihood}
\alias{get_iso_random_loglikelihood}
\title{Negative log-likelihood for a full model based on IsoAPQModelDesign object}
\usage{
get_iso_random_loglikelihood(model_design)
}
\arguments{
\item{model_design}{object of class IsoAPQModelDesign}

\item{function}{of a single parameter}
}
\description{
Negative log-likelihood for a full model based on IsoAPQModelDesign object
}
\keyword{internal}
