% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_loglikelihood_fixed.R
\name{get_neglog_fixed}
\alias{get_neglog_fixed}
\title{Negative log-likelihood for a full model}
\usage{
get_neglog_fixed(
  proteins_design,
  fixed_design,
  response,
  num_fixed_effects,
  num_proteins
)
}
\arguments{
\item{proteins_design}{design matrix for protein parameters}

\item{fixed_design}{design matrix for fixed effects}

\item{response}{response vector}

\item{num_fixed_effects}{number of fixed effects}

\item{num_proteins}{number of proteins}
}
\value{
function of a single parameter that returns negative log-likelihood
}
\description{
Negative log-likelihood for a full model
}
\keyword{internal}
