% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_log_likelihood_random.R
\name{get_neglog_random}
\alias{get_neglog_random}
\title{Negative log-likelihood for a full model}
\usage{
get_neglog_random(
  random_design,
  proteins_design,
  response,
  num_random_effects,
  num_proteins,
  random_effects_counts,
  independent_unit
)
}
\arguments{
\item{random_design}{design matrix for random effects}

\item{proteins_design}{design matrix for protein parameters}

\item{response}{response vector}

\item{num_random_effects}{number of random effects}

\item{num_proteins}{number of proteins}

\item{random_effects_counts}{counts of groups for each random effect}

\item{independent_unit}{vector of values of the independent unit}
}
\value{
function of a single parameter that returns negative log-likelihood
}
\description{
Negative log-likelihood for a full model
}
\keyword{internal}
