% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{get_protein_matrix}
\alias{get_protein_matrix}
\title{Create design matrix for proteins}
\usage{
get_protein_matrix(model_data)
}
\arguments{
\item{model_data}{data.table}
}
\description{
Create design matrix for proteins
}
\keyword{internal}
