% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoAPQModel.R
\docType{class}
\name{IsoAPQModel-class}
\alias{IsoAPQModel-class}
\alias{IsoAPQModel}
\title{S4 class to represent a nonlinear mixed effects model based on isotopic distribution}
\usage{
IsoAPQModel(model_formula, model_data, ...)
}
\arguments{
\item{model_formula}{formula that specifies the model. Special notation `..`
denotes all protein variables}

\item{model_data}{input data - data.table with columns: intensity,
iso_prob or isoDistr, sequence, rep, precursor_scan, charge_in_scan and
protein indicator columns.}

\item{additional}{parameters to the model fitting function (currently stats::optim)}
}
\value{
object of class IsoAPQModel
}
\description{
S4 class to represent a nonlinear mixed effects model based on isotopic distribution

Fit a nonlinear mixed effects model based on isotopic distribution
}
\section{Slots}{

\describe{
\item{\code{model_design}}{object of class IsoAPQModelDesign that describes
experimental design reflected in the model}

\item{\code{fitted_model_raw}}{usually list (output of stats::optim), potentially
could be a different object with raw outputs of function used to fit a model}

\item{\code{coef}}{numeric vector of model coefficient in the following order:
standard deviation of the error, parameters that describe random effects,
protein abundances, fixed effects}

\item{\code{sigma}}{standard deviation of the random error}

\item{\code{random_effects_var}}{numeric vector of estimated variances of random effects}

\item{\code{protein_abundance}}{numeric vector of estimated protein abundances}

\item{\code{fixed_effects}}{numeric vector of estimated fixed effects}

\item{\code{fitted}}{numeric vector of fitted values}

\item{\code{residuals}}{numeric vector of residuals}

\item{\code{hessian}}{hessian of the negative log-likelihood}

\item{\code{loglik}}{log-likelihood calculated at the solution}

\item{\code{message}}{message indicating if the model converged}
}}

