% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoAPQModelDesign.R
\docType{class}
\name{IsoAPQModelDesign-class}
\alias{IsoAPQModelDesign-class}
\alias{IsoAPQModelDesign}
\title{Create a description of model design for nonlinear mixed effects model}
\usage{
IsoAPQModelDesign(model_formula, model_design_list = NULL, model_data = NULL)
}
\arguments{
\item{model_design_list}{optional list of elements of the model design}

\item{inheritParams}{IsoAPQModel}
}
\value{
object of class IsoAPQModelDesign
}
\description{
Create a description of model design for nonlinear mixed effects model

Create design for nonlinear mixed effects model based on isotopic distribution
}
\section{Slots}{

\describe{
\item{\code{protein}}{protein design matrix - 0-1 matrix with a column for each protein in the dataset}

\item{\code{fixed}}{design matrix for fixed effects}

\item{\code{random}}{design matrix for random effects}

\item{\code{response}}{response - numeric vector}

\item{\code{counts}}{counts of levels of random effects}

\item{\code{formula}}{model formula}

\item{\code{independent_unit}}{character vector that specifiec independent blocks in the data}

\item{\code{num_proteins}}{number of proteins}

\item{\code{num_random_effects}}{number of random effects}

\item{\code{num_fixed_effects}}{number of fixed effects}

\item{\code{num_observations}}{number of observations}
}}

