% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_design.R
\name{parse_formula}
\alias{parse_formula}
\title{Get design matrices and response from formula and input data}
\usage{
parse_formula(model_formula, model_data)
}
\arguments{
\item{model_formula}{formula}

\item{model_data}{data.table}
}
\value{
list
}
\description{
Get design matrices and response from formula and input data
}
\keyword{internal}
